add utill.rs with feature
