+++ Cargo.toml
@@ -15,6 +15,7 @@ categories = ["asynchronous", "api-bindings", "network-programming"]
 [features]
 event = ["tokio"]
 am = ["tokio/sync", "crossbeam"]
+util = ["tokio"]
 
 [dependencies]
 ucx1-sys = { version = "0.1", path = "ucx1-sys" }
