+++ src/ucp/endpoint/stream.rs
@@ -2,7 +2,7 @@ use super::param::RequestParam;
 use super::*;
 
 impl Endpoint {
-    fn stream_send_impl(&self, buf: &[u8]) -> Result<Status<()>, Error> {
+    pub(super) fn stream_send_impl(&self, buf: &[u8]) -> Result<Status<()>, Error> {
         trace!("stream_send: endpoint={:?} len={}", self.handle, buf.len());
         unsafe extern "C" fn callback(
             request: *mut c_void,
@@ -46,7 +46,10 @@ impl Endpoint {
         }
     }
 
-    fn stream_recv_impl(&self, buf: &mut [MaybeUninit<u8>]) -> Result<Status<usize>, Error> {
+    pub(super) fn stream_recv_impl(
+        &self,
+        buf: &mut [MaybeUninit<u8>],
+    ) -> Result<Status<usize>, Error> {
         trace!("stream_recv: endpoint={:?} len={}", self.handle, buf.len());
         unsafe extern "C" fn callback(
             request: *mut c_void,
@@ -102,137 +105,9 @@ fn poll_stream(ptr: ucs_status_ptr_t) -> Poll<Result<usize, Error>> {
     }
 }
 
-use pin_project::pin_project;
-#[pin_project]
-pub struct WriteStream<'a> {
-    endpoint: &'a Endpoint,
-    #[pin]
-    request: Option<RequestHandle<Result<(), Error>>>,
-}
-
-impl Endpoint {
-    /// make write stream
-    pub fn write_stream(&self) -> WriteStream {
-        WriteStream {
-            endpoint: self,
-            request: None,
-        }
-    }
-}
-use futures::FutureExt;
-use std::task::ready;
-use tokio::io::AsyncWrite;
-use tokio::io::ReadBuf;
-
-impl<'a> AsyncWrite for WriteStream<'a> {
-    fn poll_write(
-        mut self: Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-        buf: &[u8],
-    ) -> Poll<Result<usize, std::io::Error>> {
-        if let Some(mut req) = self.as_mut().project().request.as_pin_mut() {
-            let r = match ready!(req.poll_unpin(cx)) {
-                Ok(_) => Ok(buf.len()),
-                Err(e) => Err(e.into()),
-            };
-            self.request = None;
-            Poll::Ready(r)
-        } else {
-            match self.endpoint.stream_send_impl(buf) {
-                Ok(Status::Completed(r)) => Poll::Ready(r.map(|_| buf.len()).map_err(|e| e.into())),
-                Ok(Status::Scheduled(request_handle)) => {
-                    self.request = Some(request_handle);
-                    Poll::Pending
-                }
-                Err(e) => Poll::Ready(Err(e.into())),
-            }
-        }
-    }
-
-    fn poll_flush(
-        self: Pin<&mut Self>,
-        _cx: &mut std::task::Context<'_>,
-    ) -> Poll<Result<(), std::io::Error>> {
-        todo!()
-    }
-
-    fn poll_shutdown(
-        self: Pin<&mut Self>,
-        _cx: &mut std::task::Context<'_>,
-    ) -> Poll<Result<(), std::io::Error>> {
-        todo!()
-    }
-}
-
-#[pin_project]
-pub struct ReadStream<'a> {
-    endpoint: &'a Endpoint,
-    #[pin]
-    request: Option<RequestHandle<Result<usize, Error>>>,
-}
-
-impl Endpoint {
-    /// make read stream
-    pub fn read_stream(&self) -> ReadStream {
-        ReadStream {
-            endpoint: self,
-            request: None,
-        }
-    }
-}
-
-use tokio::io::AsyncRead;
-impl<'a> AsyncRead for ReadStream<'a> {
-    fn poll_read(
-        mut self: Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-        out_buf: &mut ReadBuf<'_>,
-    ) -> Poll<Result<(), std::io::Error>> {
-        if let Some(mut req) = self.as_mut().project().request.as_pin_mut() {
-            let r = match ready!(req.poll_unpin(cx)) {
-                Ok(n) => {
-                    // Safety: The buffer was filled by the recv operation.
-                    unsafe {
-                        out_buf.assume_init(n);
-                        out_buf.advance(n);
-                    }
-                    Ok(())
-                }
-                Err(e) => Err(e.into()),
-            };
-            self.request = None;
-            Poll::Ready(r)
-        } else {
-            let buf = unsafe { out_buf.unfilled_mut() };
-            match self.endpoint.stream_recv_impl(buf) {
-                Ok(Status::Completed(n_result)) => {
-                    match n_result {
-                        Ok(n) => {
-                            // Safety: The buffer was filled by the recv operation.
-                            unsafe {
-                                out_buf.assume_init(n);
-                                out_buf.advance(n);
-                            }
-                            Poll::Ready(Ok(()))
-                        }
-                        Err(e) => Poll::Ready(Err(e.into())),
-                    }
-                }
-                Ok(Status::Scheduled(request_handle)) => {
-                    self.request = Some(request_handle);
-                    cx.waker().wake_by_ref();
-                    Poll::Pending
-                }
-                Err(e) => Poll::Ready(Err(e.into())),
-            }
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use tokio::io::{AsyncReadExt, AsyncWriteExt};
     #[test_log::test]
     fn stream() {
         for i in 0..20_usize {
@@ -289,22 +164,6 @@ mod tests {
             }
         );
 
-        tokio::join!(
-            async {
-                // send
-                let buf = vec![42u8; msg_size];
-                endpoint2.write_stream().write_all(&buf).await.unwrap();
-                println!("write_stream");
-            },
-            async {
-                // recv
-                let mut buf = vec![0u8; msg_size];
-                endpoint1.read_stream().read_exact(&mut buf).await.unwrap();
-                assert_eq!(buf, vec![42u8; msg_size]);
-                println!("read_stream");
-            }
-        );
-
         println!("status {:?}", endpoint2.get_status());
         assert_eq!(endpoint1.get_rc(), (1, 1));
         assert_eq!(endpoint2.get_rc(), (1, 1));
@@ -314,68 +173,5 @@ mod tests {
         assert_eq!(endpoint2.get_rc(), (1, 1));
         assert_eq!(endpoint2.close(true).await, Ok(()));
         assert_eq!(endpoint2.get_rc(), (1, 0));
-    }
-
-    #[test_log::test]
-    fn stream_send_recv_various_contents_and_sizes() {
-        spawn_thread!(_stream_send_recv_various_contents_and_sizes())
-            .join()
-            .unwrap();
-    }
-
-    async fn _stream_send_recv_various_contents_and_sizes() {
-        let context1 = Context::new().unwrap();
-        let worker1 = context1.create_worker().unwrap();
-        let context2 = Context::new().unwrap();
-        let worker2 = context2.create_worker().unwrap();
-        tokio::task::spawn_local(worker1.clone().polling());
-        tokio::task::spawn_local(worker2.clone().polling());
-
-        // connect with each other
-        let mut listener = worker1
-            .create_listener("0.0.0.0:0".parse().unwrap())
-            .unwrap();
-        let listen_port = listener.socket_addr().unwrap().port();
-        let mut addr: std::net::SocketAddr = "127.0.0.1:0".parse().unwrap();
-        addr.set_port(listen_port);
-
-        let (endpoint1, endpoint2) = tokio::join!(
-            async {
-                let conn1 = listener.next().await;
-                worker1.accept(conn1).await.unwrap()
-            },
-            async { worker2.connect_socket(addr).await.unwrap() },
-        );
-
-        // Test cases: (data, repeat count)
-        let test_cases = vec![
-            (vec![], 1),
-            (vec![0u8], 10),
-            (vec![1, 2, 3, 4, 5], 5),
-            ((0..128).collect::<Vec<u8>>(), 3),
-            ((0..1024).map(|x| (x % 256) as u8).collect::<Vec<u8>>(), 2),
-            ((0..4096).map(|x| (x % 256) as u8).collect::<Vec<u8>>(), 1),
-        ];
-        for (data, repeat) in test_cases {
-            for _ in 0..repeat {
-                // send
-                let send_buf = data.clone();
-                let recv_len = send_buf.len();
-                let mut recv_buf = vec![0u8; recv_len];
-                tokio::join!(
-                    async {
-                        endpoint2.write_stream().write_all(&send_buf).await.unwrap();
-                    },
-                    async {
-                        endpoint1
-                            .read_stream()
-                            .read_exact(&mut recv_buf)
-                            .await
-                            .unwrap();
-                        assert_eq!(recv_buf, send_buf, "data mismatch for len={}", recv_len);
-                    }
-                );
-            }
-        }
     }
 }
